Replace ffi with definitions from libc
The C data types vary slightly between architectures, so it is best to
use the definitions from the crate. This fixes a compile error on ARM
where c_char is u8, so it cannot be used interchangeably with c_schar
which is always i8.

This changes the public interface for ptsname() to return c_char rather
than c_schar, but those are the same on x86 and it wouldn't have
compiled on other architectures, so it is not an API break.

  --> src/fork/mod.rs:35:60
   |
35 | Ok(name) => Fork::from_pts(name),
   |                            ^^^^ expected u8, found i8
   |
   = note: expected type `*const u8`
              found type `*const i8`